Improve code quality and documentation
- Add rust-version = 1.73 and exclude directive to Cargo.toml
- Update README.md to clarify arbitrary size support
- Add comprehensive crate-level documentation to lib.rs
- Add Debug derives to GpuFft and SizeCache
- Fix clippy warnings in benchmark.rs and fft.rs
- Add Default implementations for rivals types
- Add clippy allows to rivals module (experimental code)
- Add CHANGELOG.md for version tracking

+++ src/benchmark.rs
@@ -70,7 +70,7 @@ pub fn benchmark_gpu_only(
         })
         .collect();
 
-    let mut all_raw_data = Vec::with_capacity((n * 2 * batch_size as usize) as usize);
+    let mut all_raw_data = Vec::with_capacity(n * 2 * batch_size);
     for input in &inputs {
         let raw = gpu_fft.prepare_input_data(input, false);
         all_raw_data.extend_from_slice(&raw);
